optimize the entire compiler by using const ref for passing strings, vectors, etc. also improve the compile times of the compiler alittle bit, im still working at it.

+++ source/language/lists.hpp
@@ -9,10 +9,11 @@
 #ifndef lists_hpp
 #define lists_hpp
 
+#include <cstdint>
 #include <vector>
 #include <string>
 
-using nat = size_t;
+using nat = int_fast64_t;
 
 // language info
 extern const std::string language_name;
@@ -20,7 +21,6 @@ extern const std::string language_version;
 
 extern const std::string compiler_name;
 extern const std::string compiler_version;
-
 
 extern nat spaces_count_for_indent;
 extern nat max_expression_depth;
